refactor: move codex related from `protocol` to `server`.
