+++ internal/server/codex_responses.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
+	"github.com/openai/openai-go/v3/packages/param"
 	"github.com/openai/openai-go/v3/packages/ssestream"
 	"github.com/openai/openai-go/v3/responses"
 	"github.com/sirupsen/logrus"
@@ -20,6 +21,77 @@ import (
 	streamhandler "github.com/tingly-dev/tingly-box/internal/protocol/stream"
 	"github.com/tingly-dev/tingly-box/internal/typ"
 )
+
+// ConvertResponsesToChatGPTFormat converts OpenAI Responses API params to ChatGPT backend API format.
+func ConvertResponsesToChatGPTFormat(params responses.ResponseNewParams, provider *typ.Provider) *request.ChatGPTBackendRequest {
+	req := &request.ChatGPTBackendRequest{
+		Model:      string(params.Model),
+		Stream:     true,
+		Tools:      []interface{}{},
+		ToolChoice: "auto",
+		Store:      false,
+		Include:    []string{},
+	}
+
+	// Add instructions if present, otherwise use default
+	// ChatGPT backend API requires instructions to be present
+	if !param.IsOmitted(params.Instructions) {
+		req.Instructions = params.Instructions.Value
+	} else {
+		req.Instructions = "You are a helpful AI assistant."
+	}
+
+	// Convert input to ChatGPT backend API format
+	if !param.IsOmitted(params.Input.OfInputItemList) {
+		req.Input = request.ConvertResponseInputToChatGPTFormat(params.Input.OfInputItemList)
+	}
+
+	// Convert tools to ChatGPT backend API format
+	if !param.IsOmitted(params.Tools) && len(params.Tools) > 0 {
+		req.Tools = request.ConvertResponseToolsToChatGPTFormat(params.Tools)
+	}
+
+	// Convert tool_choice if present
+	if !param.IsOmitted(params.ToolChoice) {
+		req.ToolChoice = request.ConvertResponseToolChoiceToChatGPTFormat(params.ToolChoice)
+	}
+
+	// Copy other fields if present
+	// Note: Codex OAuth providers do not support max_tokens, max_completion_tokens, temperature, or top_p
+	if !param.IsOmitted(params.MaxOutputTokens) {
+		// Skip for Codex OAuth providers
+		if !IsCodexOAuthProvider(provider) {
+			if request.RequiresMaxCompletionTokens(string(params.Model)) {
+				req.MaxCompletion = int(params.MaxOutputTokens.Value)
+			} else {
+				req.MaxTokens = int(params.MaxOutputTokens.Value)
+			}
+		}
+	}
+	if !param.IsOmitted(params.Temperature) {
+		// Skip for Codex OAuth providers
+		if !IsCodexOAuthProvider(provider) {
+			req.Temperature = params.Temperature.Value
+		}
+	}
+	if !param.IsOmitted(params.TopP) {
+		// Skip for Codex OAuth providers
+		if !IsCodexOAuthProvider(provider) {
+			req.TopP = params.TopP.Value
+		}
+	}
+
+	return req
+}
+
+// IsCodexOAuthProvider checks if the provider is a Codex OAuth provider.
+// Codex OAuth providers do not support max_tokens or max_completion_tokens parameters.
+func IsCodexOAuthProvider(provider *typ.Provider) bool {
+	if provider == nil || provider.OAuthDetail == nil {
+		return false
+	}
+	return provider.OAuthDetail.ProviderType == "codex"
+}
 
 // forwardChatGPTBackendRequest forwards a request to ChatGPT backend API using the correct format
 // Reference: https://github.com/SamSaffron/term-llm/blob/main/internal/llm/chatgpt.go
@@ -185,7 +257,7 @@ func (s *Server) handleChatGPTBackendStreamingRequest(c *gin.Context, provider *
 // makeChatGPTBackendRequest creates and executes an HTTP request to ChatGPT backend API
 func (s *Server) makeChatGPTBackendRequest(wrapper *client.OpenAIClient, provider *typ.Provider, params responses.ResponseNewParams) (*http.Response, context.CancelFunc, error) {
 	// Convert OpenAI Responses API params to ChatGPT backend API format using protocol handler
-	chatGPTReq := request.ConvertResponsesToChatGPTFormat(params, provider)
+	chatGPTReq := ConvertResponsesToChatGPTFormat(params, provider)
 
 	bodyBytes, err := json.Marshal(chatGPTReq)
 	if err != nil {
